# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IndianTradeMart API
  version: 1.0.0
  description: Express API surface for the IndianTradeMart platform.
servers:
  - url: http://localhost:3001
    description: Local dev server
  - url: / 
    description: Relative (proxy)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SupportTicketCreate:
      type: object
      required: [subject, description]
      properties:
        subject:
          type: string
        description:
          type: string
        category:
          type: string
        priority:
          type: string
        status:
          type: string
        vendor_id:
          type: string
        buyer_id:
          type: string
        attachments:
          type: array
          items:
            type: object
    SupportMessageCreate:
      type: object
      required: [message]
      properties:
        message:
          type: string
        sender_type:
          type: string
        sender_id:
          type: string
...
paths:
...
  /api/support/tickets:
    get:
      summary: List tickets (filters)
      tags: [Support]
      responses:
        '200':
          description: Tickets
    post:
      summary: Create ticket
      tags: [Support]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportTicketCreate'
      responses:
        '201':
          description: Created
  /api/support/tickets/{id}:
    get:
      summary: Get ticket
      tags: [Support]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket
    patch:
      summary: Update ticket fields
      tags: [Support]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
  /api/support/tickets/{id}/messages:
    get:
      summary: Get ticket messages
      tags: [Support]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages
    post:
      summary: Post ticket message
      tags: [Support]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportMessageCreate'
      responses:
        '200':
          description: Created
  /api/support/tickets/{id}/status:
    put:
      summary: Update ticket status
      tags: [Support]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        '200':
          description: Updated
...
  /api/support/vendor/{vendorId}:
    get:
      summary: Vendor tickets
      tags: [Support]
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tickets
  /api/support/buyer/{buyerId}:
    get:
      summary: Buyer tickets
      tags: [Support]
      parameters:
        - name: buyerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tickets
...